pyproject.toml
[project]
name = "chrona"
version = "0.1.0"
description = "Smart Schedule Optimizer powered by Google Gemini AI"
authors = []
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.28.0",
    "google-genai>=0.3.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "typing-extensions>=4.5.0",
    "python-dotenv>=1.0.0",
    "google-auth>=2.17.0",
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.1.0",
    "google-api-python-client>=2.88.0",
]
